Fix: Error emitters must have at least one emitter
Per https://nodejs.org/api/events.html#events_error_events - when
we receive an error from the server we raise an error event to
be used for monitoring and logging purposes.

If node.js treats the error event specially and, if there is no
registered listener, treats the event as an unhandled exception
just to mess with you.